Fix Color control HSV value

+++ examples/lighting-app/lighting-common/include/ColorFormat.h
@@ -29,7 +29,7 @@ struct RgbColor_t
 
 struct HsvColor_t
 {
-    uint8_t h;
+    uint16_t h;
     uint8_t s;
     uint8_t v;
 };
